refactor out resources
