+++ app.py
@@ -1,20 +1,27 @@
 import os
 from os import urandom
 from pymongo.uri_parser import parse_uri
-from flask import Flask, request, jsonify
+from flask import Flask, jsonify, request
 from flask_mongoengine import MongoEngine, MongoEngineSessionInterface
 from flask_restful import Api
 from flask_cors import CORS
 from flask_jwt_extended import JWTManager, create_access_token
-from mini_gplus.resources import Users, Posts, Comments, NestedComments, Circles, Circle, CircleMember, \
-    Followings, Following, Profile, UserResource, Reactions, Reaction, MyAvatar, Me, MyProfilePic, Post, \
-    Notifications, Home, PostMedia
-from mini_gplus.models import User
+from mini_gplus.daos.user import check_user, create_user
+from mini_gplus.resources.me import MyAvatar, MyProfilePic, Me
+from mini_gplus.resources.users import Users, User
+from mini_gplus.resources.posts import Profile, Home, PostMedia, Posts, Post
+from mini_gplus.resources.comments import NestedComments, Comments
+from mini_gplus.resources.reactions import Reactions, Reaction
+from mini_gplus.resources.circles import Circles, CircleMember, Circle
+from mini_gplus.resources.followings import Followings, Following
+from mini_gplus.resources.notifications import Notifications
 
 
 app = Flask(__name__)
+
 app.secret_key = urandom(24)
 
+# database
 mongodb_uri = os.environ['MONGODB_URI']
 mongodb_db = parse_uri(mongodb_uri)['database']
 app.config['MONGODB_SETTINGS'] = {
@@ -26,14 +33,18 @@ app.session_interface = MongoEngineSessionInterface(db)
 
 app.config['MAX_CONTENT_LENGTH'] = 40 * 1024 * 1024  # 40MB max upload size
 
-##################
-# Authentication #
-##################
+# jwt
 app.config['JWT_SECRET_KEY'] = os.environ['JWT_SECRET_KEY']
 app.config['JWT_ACCESS_TOKEN_EXPIRES'] = False
 JWTManager(app)
 
+app.config['BUNDLE_ERRORS'] = True
 
+# cors
+CORS(app, resources={r"/api/*": {"origins": "*"}})
+
+
+# auth
 @app.route('/api/signIn', methods=['POST'])
 def sign_in():
     """
@@ -47,7 +58,7 @@ def sign_in():
         return jsonify({"message": {"id": "id is required"}}), 400
     if not password:
         return jsonify({"message": {"password": "password is required"}}), 400
-    user_checked = User.check(user_id, password)
+    user_checked = check_user(user_id, password)
     if not user_checked:
         return jsonify({"message": "invalid id or password"}), 401
     access_token = create_access_token(identity=user_id)
@@ -67,19 +78,14 @@ def sign_up():
         return jsonify({"message": {"id": "id is required"}}), 400
     if not password:
         return jsonify({"message": {"password": "password is required"}}), 400
-    successful = User.create(user_id, password)
+    successful = create_user(user_id, password)
     if successful:
         return {'id': user_id}, 201
     else:
         return {'msg': f'ID {user_id} is already taken'}, 409
 
 
-########
-# APIs #
-########
-app.config['BUNDLE_ERRORS'] = True
-CORS(app, resources={r"/api/*": {"origins": "*"}})
-
+# api
 api = Api(app, errors={
     'UnauthorizedAccess': {
         'status': 401,
@@ -95,8 +101,9 @@ api = Api(app, errors={
 api.add_resource(MyAvatar, '/api/me/avatar')
 api.add_resource(MyProfilePic, '/api/me/profilePic/<string:user_profile_pic>')
 api.add_resource(Me, '/api/me')
+
 api.add_resource(Users, '/api/users')
-api.add_resource(UserResource, '/api/user/<string:user_id>')
+api.add_resource(User, '/api/user/<string:user_id>')
 
 api.add_resource(Profile, '/api/profile/<string:profile_user_id>')
 api.add_resource(Home, '/api/home')
