+++ mini_gplus/models/__init__.py
@@ -4,3 +4,4 @@ from .notification import Notification, NotifyingAction
 from .reaction import Reaction
 from .comment import Comment
 from .post import Post
+from .media import Media
